refactor: util패키지 삭제 및 generator를 domain으로 이동
