+++ src/test/java/domain/LinesTest.java
@@ -7,8 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import util.ExistFootholdGenerator;
-import util.NonExistFootholdGenerator;
+
+import domain.generator.ExistFootholdGenerator;
+import domain.generator.NonExistFootholdGenerator;
 
 public class LinesTest {
